+ Qt version update
git-svn-id: svn://svn.code.sf.net/p/mediainfo/code/MediaInfo/trunk@3337 45c5ae6f-87cc-4fd0-92ee-e4f023fd80da

+++ Source/Common/Core.cpp
@@ -311,6 +311,6 @@ String Core::Parameters () {
     return MediaInfoNameSpace::MediaInfo::Option_Static(_T("Info_Parameters_CSV"));
 }
 
-int Core::State_Get() {
+size_t Core::State_Get() {
     return MI->State_Get();
 }
